+++ crates/aviutl2/src/input/bridge.rs
@@ -138,6 +138,7 @@ pub unsafe fn initialize_plugin_c_unwind<T: InputSingleton>(version: u32) -> boo
 }
 
 pub(crate) fn initialize_plugin<T: InputSingleton>(version: u32) -> AnyResult<()> {
+    crate::common::ensure_minimum_aviutl2_version(version.into())?;
     let plugin_state = T::__get_singleton_state();
     let info = crate::common::AviUtl2Info {
         version: version.into(),