+++ crates/aviutl2/src/module/bridge.rs
@@ -71,6 +71,7 @@ pub unsafe fn initialize_plugin_c_unwind<T: ScriptModuleSingleton>(version: u32)
 }
 
 pub(crate) fn initialize_plugin<T: ScriptModuleSingleton>(version: u32) -> AnyResult<()> {
+    crate::common::ensure_minimum_aviutl2_version(version.into())?;
     let plugin_state = T::__get_singleton_state();
     let info = crate::common::AviUtl2Info {
         version: version.into(),