+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- **Breaking**: beta30未満を明示的に弾くように
 - module: i32・f64以外にもFromScriptModuleParamを実装
 - ログを自動で折り返すように
 - デフォルトでpanicしたときにunwindするように